Embed the managed part (C#) into the C++/CLI wrapper
The major difficulty I had to overcome was that as soon as the
wrapper is loaded and some code referencing the C# part is executed,
the managed dll is automatically looked for before I can set an
assembly resolver, which triggers a FileNotFound exception if the dll
is not found in the working folder.

Since I originally only wanted to put the C# dll into a subfolder,
an app.config with a <probing> directive would have sufficed,
but I did not want another config file poluting the plugins directory.

The trick was simply to call an initializer function, not related to
any C# reference, hence not triggering the auto-loading of the
dependent dll.

Once this was in place and working, going one step further and
actually embedding and loading the C# managed dll from a resource
was easy :)