Embed the C++/CLI wrapper into the native plugin
The compiled wrapper dll is embedded into the native plugin as a
resource (see main.rc).

A hook into the dll delayed loading process is then set so that when
the wrapper is needed, the correct module can be loaded based on the
required dll name.
(Which is hardcoded into the native plugin based on the wrapper.lib
file, if I'm not mistaken. In any case, using LoadLibrary to load the
wrapper renamed with a random name does not work. Without the hook,
the file HAS to have its original name...)

Currently, the extracted resource needs to be loaded via an explicit
call to InitializeEmbeddedManagedWrapper(), but this might change
in the future...
(This is not very practical because this call must not be made from
the DllMain entry point (see DllMain loader lock errors), but has to
be made before any code from the wrapper is executed :/)