Automatically load the embedded wrapper from the native plugin
No need for a weird initialization function anymore!
Everything is automatic: extraction of the resource to the temp
folder, loading of the extracted file, and resolution of the dll
when required via the delayed dll loading process!

Since this is lazily done when code from the wrapper is required to
be executed, this might still require a critical section or something
similar to properly handle an initialization from multiple threads
though...